+++ src/database/models/14-producto.js
@@ -50,7 +50,7 @@ module.exports = (sequelize, DataTypes) => {
     },
 
     categoria_id: {
-      type: DataTypes.TINYINT.UNSIGNED,
+      type: DataTypes.SMALLINT.UNSIGNED,
       allowNull: false,
       defaultValue: 1,
       references: {
